refactor cslb extraction
